Separate out logic to different files and use by imports
